Fix search score type cast and update equipment TSV
Cast similarity score to float8 for sqlx compatibility.
Fixed TSV column alignment for antenna/key/mic rows.

Bump version to 0.4.58

+++ src/db/equipment.rs
@@ -78,7 +78,7 @@ pub async fn search_equipment(
                    similarity(name, $1),
                    similarity(manufacturer || ' ' || name, $1),
                    similarity(array_to_string(aliases, ' '), $1)
-               ) AS score,
+               )::float8 AS score,
                CASE
                    WHEN similarity(name, $1) >= GREATEST(
                        similarity(manufacturer || ' ' || name, $1),
